Add regression API
